\name{interferenceCI-package}
\alias{interferenceCI-package}
\docType{package}
\title{Exact confidence intervals in the presence of interference}
\description{Computes large sample confidence intervals of Liu and
  Hudgens (2014), conservative exact confidence intervals of Tchetgen
  Tchetgen and VanderWeele (2012), and exact confidence intervals of
  Rigdon and Hudgens (2014) for treatment effects on a binary outcome in two-stage randomized experiments with interference.}
\details{
\tabular{ll}{
Package: \tab interferenceCI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-16\cr
License: \tab GPL-3\cr
}
}

\author{Joseph Rigdon \email{jrigdon@bios.unc.edu}}

\references{Hudgens, M.G. and Halloran, M.E. ``Toward causal inference
  with interference.''  Journal of the
  American Statistical Association 2008 103:832-842.

  Tchetgen Tchetgen, E. and VanderWeele, T.J. ``On causal
  inference in the presence of interference.''  Statistical Methods in
  Medical Research 2012 21:55-75.

  Liu, L. and Hudgens, M.G. ``Large sample randomization inference of
  causal effects in the presence of interference.''  Journal of the
  American Statistical Association 2014 109, 288-301.

  Rigdon, J. and Hudgens, M.G. ``Exact confidence intervals in
  the presence of interference.'' Submitted to Statistics and Probability Letters 2014.}